moved toolkit code to `compile/toolkit`
