+++ backend/internal/compile/toolkit/toolkit_dev.go
@@ -1,6 +1,6 @@
 //go:build !prod
 
-package compilerServer
+package toolkit
 
 import (
 	"io/fs"
@@ -11,12 +11,12 @@ import (
 	"robinplatform.dev/internal/log"
 )
 
-var toolkitFS fs.FS
+var ToolkitFS fs.FS
 
 var initToolkit = func() {
 	_, filename, _, _ := runtime.Caller(0)
 	toolkitPath := filepath.Clean(filepath.Join(filepath.Dir(filename), "..", "..", "..", "toolkit"))
-	toolkitFS = os.DirFS(toolkitPath)
+	ToolkitFS = os.DirFS(toolkitPath)
 	logger.Warn("Detected dev mode, using local toolkit", log.Ctx{
 		"path": toolkitPath,
 	})
